Lights on spaceship

+++ src/spaceship.h
@@ -1,4 +1,4 @@
-extern double spaceship_altitude;
+extern float spaceship_altitude;
 extern bool spaceship_flying;
 extern double spaceship_height;
